+++ app/models/randomization_scheme.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+# Defines a method for randomization, either Permuted-Block or Minimization
 class RandomizationScheme < ActiveRecord::Base
   # Triggers
   after_create :create_default_block_size_multipliers
@@ -15,21 +16,23 @@ class RandomizationScheme < ActiveRecord::Base
   scope :published, -> { where published: true }
 
   # Model Validation
-  validates_presence_of :name, :user_id, :project_id
-  validates_uniqueness_of :name, case_sensitive: false, scope: [:deleted, :project_id]
-  validates_numericality_of :randomization_goal, greater_than_or_equal_to: 0, only_integer: true
-  validates_numericality_of :chance_of_random_treatment_arm_selection, greater_than_or_equal_to: 0, less_than_or_equal_to: 100, only_integer: true
+  validates :name, :user_id, :project_id, presence: true
+  validates :name, uniqueness: { case_sensitive: false, scope: [:deleted, :project_id] }
+  validates :randomization_goal,
+            numericality: { greater_than_or_equal_to: 0, only_integer: true }
+  validates :chance_of_random_treatment_arm_selection,
+            numericality: { greater_than_or_equal_to: 0, less_than_or_equal_to: 100, only_integer: true }
 
   # Model Relationships
   belongs_to :user
   belongs_to :project
   belongs_to :variable
-  has_many :block_size_multipliers, -> { where(deleted: false).order(:value) }
-  has_many :lists,                  -> { where deleted: false }
-  has_many :randomizations,         -> { where deleted: false }
-  has_many :stratification_factors, -> { where deleted: false }
+  has_many :block_size_multipliers,        -> { where(deleted: false).order(:value) }
+  has_many :lists,                         -> { where deleted: false }
+  has_many :randomizations,                -> { where deleted: false }
+  has_many :stratification_factors,        -> { where deleted: false }
   has_many :stratification_factor_options, -> { where deleted: false }
-  has_many :treatment_arms,         -> { where(deleted: false).order(:name) }
+  has_many :treatment_arms,                -> { where(deleted: false).order(:name) }
 
   # Model Methods
 
@@ -38,10 +41,10 @@ class RandomizationScheme < ActiveRecord::Base
   end
 
   def generate_lists!(current_user)
-    return false if self.has_randomized_subjects?
-    self.randomizations.destroy_all
-    self.lists.destroy_all
-    self.add_missing_lists!(current_user)
+    return false if randomized_subjects?
+    randomizations.destroy_all
+    lists.destroy_all
+    add_missing_lists!(current_user)
     true
   end
 
@@ -62,35 +65,44 @@ class RandomizationScheme < ActiveRecord::Base
   end
 
   def minimum_block_size
-    self.treatment_arms.pluck(:allocation).sum
+    treatment_arms.pluck(:allocation).sum
   end
 
   def randomization_requirements
-    # "By checking this box I attest that I have personally entered all of the available data recorded and reviewed for completeness and accuracy. All information entered by me is correct to the best of my knowledge."
+    # "By checking this box I attest that I have personally entered all of the
+    #   available data recorded and reviewed for completeness and accuracy. All
+    #   information entered by me is correct to the best of my knowledge."
   end
 
-  def has_randomized_subjects?
-    self.randomizations.where.not(subject_id: nil).count > 0
+  def randomized_subjects?
+    randomizations.where.not(subject_id: nil).count > 0
   end
 
   def algorithm_name
-    if algorithm = ALGORITHMS.select{|name, value| value == self.algorithm}.first
+    algorithm = ALGORITHMS.find { |_name, value| value == algorithm }
+    if algorithm
       algorithm[0]
     else
-      "No Algorithm Selected"
+      'No Algorithm Selected'
     end
   end
 
   def permuted_block?
-    self.algorithm == 'permuted-block'
+    algorithm == 'permuted-block'
   end
 
   def minimization?
-    self.algorithm == 'minimization'
+    algorithm == 'minimization'
   end
 
   def randomization_error_message
     RandomizationAlgorithm.for(self).randomization_error_message
+  end
+
+  def stratification_factors_with_calculation
+    stratification_factors
+      .where(stratifies_by_site: false)
+      .where.not(calculation: ['', nil])
   end
 
   private