+++ db/migrate/20160212145449_add_calculation_to_stratification_factors.rb
@@ -1,0 +1,5 @@
+class AddCalculationToStratificationFactors < ActiveRecord::Migration
+  def change
+    add_column :stratification_factors, :calculation, :text
+  end
+end